Improve TSNE calculations and button behaviour in frontend
The commit refactors the calculation methods in Tsne.tsne.go to improve their readability and performance. The changes include simplifying gradient calculation, refactoring variable declaration and calculation, and optimizing sum calculation.

On the frontend side, this commit modifies the event-handler for the Generate 3D Graph button. It now disables the button while it's generating the graph and re-enables it when the task is finished to provide better user feedback. The number of iterations for generating the 3D Graph has also been reduced from 20 to 10, which should improve response time.

Methods CreateTSNE and GetTSNEData in Server/routes.go are now logged for more effortless tracking.