Refactored collection and index handling in VreeDB
Refactoring of the VreeDB project has been done with changes in collection.go, routes.go, and Index.go. It involved renaming Mut to mut within the index handling, modifying file opening mode of Indexes from Open to Create and altering the payload loading order. Additionally, the requirement of ApiKey field check in the IndexCreator of Server routes.go was omitted.