Refactor Filter.go to simplify type cases
The Filter.go file has been refactored to simplify the switch cases for different types. The separate handling of types such as int, int64, float32, float64 and string has been removed and replaced with handling of the []interface{} type. This streamlined the filter function and minimizes redundancy in the code.