Add search threads to ArgsParser and SearchUnit
A new field "SearchThreads" has been added to the ArgsParser struct to allow configuration of the number of search threads. Concurrent searching has been implemented in the Start method of the SearchUnit by iterating over the "SearchThreads" count and starting a new goroutine for each one. This allows nearby neighbor searches to be performed in parallel for improved performance.