Refactor code for vector collection and thread synchronization
This commit refines the process of saving a vector to a file system, with changes ensuring the action only occurs if the vector is new. It also improves usage of waitgroups in the SearchUnit model, enhancing thread synchronization. Amendment introduced in vector initialization ensures new and non-indexed vectors have a save position value set to -1 by default.