Refactor logging message construction
The logic for constructing a logging message has been moved from the `Log` function to the case where a message is received from the channel in the logger service loop. The change simplifies the `Log` function to only wrap the message and level into a `LogMessage` struct and send to the channel, while the formatting process with time stamp is now handled in the service loop.