Switch mutex lock type and implement SetLocalDiaSpace method
Changed Mutex Lock to RLock in the DeleteWatcher method to allow concurrent read. Also, properly implemented the SetLocalDiaSpace method in collection.go which sets the diagonal space of the Collection in local variables. It now updates the max and min vectors, calculates their difference, and calculates the diagonal length of the Collection.