Refactor logging and mutex placement
The adjustments made refactor the synchronizing of threads and the logging system. Mutex locks have been repositioned in the DeleteMarkedVectors() function to ensure proper synchronization of threads. In addition, the logging level has been converted to uppercase in logger.go and the log messages have been corrected from "ERROR" to "INFO" in bootup.go to provide more accurate information about the system state.