Add delete operation with filter functionality
This update introduces two key changes. First, a SerialDelete function was added to the Collection package to delete vectors based on provided filters. Second, a DeletePointWithFilter function was implemented in the Server package to handle HTTP requests for deletions with filters. Updating of the DeletePoint struct in the Server package was also required to accommodate this new feature.