Add epsilon and improve float comparison in NeuralNet and SVM
The constant epsilon was introduced due to the inability to compare two floats directly because of how they are represented in memory. Epsilon is now used in our SVM comparison to enhance the precision of the operation. This also involved adding a comment to identify the point at which the Alpha is updated.